Add files via upload
The initial focus for this project was on investing and the power of compound interest. I developed an application that allows users to see how their investments will grow over time.
In looking at the written code, the calculation was well written to deliver the expected output. 
The code could be enhanced greatly by developing more classes and a private/public structure to expand on the bank’s organization. 
The most challenging part of the project was getting the expected output organized to a readable layout. Utilizing social message boards to help collaborate needed results when presented with roadblocks is essential to a supporting network.  
Skills in formulating calculations will be beneficial to other applications of similar results along with projects that require such user data input. 
The project was made maintainable for its readable code and good use of spacing. By using inline comments, the project could be easily understood if intergraded into another project introduced to another group.